Initial support for parsing BitcoinD Scripts
As part of building full support for a compatible Script interpreter,
I am adding tests that run the full bitcoin script test vectors. Those
test vectors are in a text format that have specific parsing
requirements.
This commit is the first step on that road. Only a few test vectors pass
at the moment.