Completed refactor of decoupled ScriptSig & ScriptPubkey
- When creating transaction input and output scripts it is important
to mainting the flexibility that bitcoin was built with. The concept
of a "standard script" is going away. This refactor has for the most
part removed the limits on all Transactions bein P2PKH transactions by
allowing the developer to have arbitrary locking/unlocking scripts
associated with their inputs and outputs.